refactor(rust): extract ooxml-common crate for shared color transforms
Adds a new workspace member `packages/ooxml-common` that hosts the
OOXML color-transform helpers previously duplicated between the docx
and pptx parsers (~200 lines apiece).

Shared
- apply_color_transforms (lumMod/lumOff, satMod/satOff, hueMod/hueOff,
  shade, tint, alpha + alphaMod/alphaModFix/alphaOff)
- srgb_to_linear / linear_to_srgb
- rgb_to_hls / hls_to_rgb (Python-colorsys-style HLS conversion)

The two existing tint divergences are kept as a single TintMode flag:
- TintMode::WordLiteral — `result = val·input + (1-val)·white` per the
  literal ECMA-376 §20.1.2.3.34 reading. Used by docx; this is what
  matches Word's render of resume / cover templates.
- TintMode::PowerPointLinear — `lerp(input, white, val)` in linear sRGB.
  Used by pptx; verified against PDF exports of SmartArt accent recolors.

Both consumer parsers now keep just a thin one-line wrapper that picks
the appropriate TintMode and forwards to the shared helper. Net diff:
docx parser loses ~150 lines, pptx parser loses ~140 lines, ooxml-common
adds 200 (with a single source of truth and per-app comments).

VRT
- docx: all sample match% identical to before; demo/sample-1 100%.
- pptx: 69/69 passing; demo/sample-1 100%.
