+++ packages/docx/parser/Cargo.toml
@@ -14,3 +14,4 @@ zip = { workspace = true }
 roxmltree = { workspace = true }
 base64 = { workspace = true }
 console_error_panic_hook = { workspace = true }
+ooxml-common = { workspace = true }
