+++ packages/ooxml-common/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ooxml-common"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["rlib"]
+
+[dependencies]
+roxmltree = { workspace = true }
